requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.2
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
faiss-cpu==1.11.0
fastapi==0.116.1
filelock==3.18.0
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-text-splitters==0.3.8
langsmith==0.4.5
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.5
numpy==1.26.4
openai==1.95.1
opencv-python==4.8.1.78
orjson==3.10.18
packaging==24.2
pandas==2.3.1
pillow==11.3.0
pip==25.1
propcache==0.3.2
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
safetensors==0.5.3
scipy==1.16.0
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.2
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.67.1
transformers==4.31.0
typing_extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
ultralytics==8.3.165
ultralytics-thop==2.0.14
urllib3==2.5.0
uvicorn==0.35.0
wheel==0.45.1
yarl==1.20.1
zstandard==0.23.0